+++ script/make-run-server-stdio-java.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+cd "$(dirname "$0")/.."
+make run-server-stdio-java
+